reflect recent PRs
Updated the README to reflect the changes added from Greg's recent pull
requests.  The -g flag or --getalertstyle is the most noticeable
improvement as it allows you to get the current alert style for any
app.

I also updated the alert style syntax to the plural form of the alert
style.

Advised that changing the _SYSTEM_CENTER_ items may not be the best
idea, but it seems to work.